src/dashboard/requirements.txt
python-dotenv~=1.0.0
requests~=2.28.2
dash
pandas
dash_bootstrap_components~=1.4.1
